Refactor of hittable abstraction + diffuse shaders

+++ src/core/ray.c
@@ -6,7 +6,7 @@
 /*   By: kdaniely <kdaniely@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/07 14:16:55 by kdaniely          #+#    #+#             */
-/*   Updated: 2024/04/16 02:40:05 by kdaniely         ###   ########.fr       */
+/*   Updated: 2024/05/10 17:33:12 by kdaniely         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	new_ray_from_to(t_ray *self, t_point3 from, t_point3 to)
 	self->direction = dir;
 }
 
-t_point3	ray_at(t_ray *r, float t)
+t_point3	ray_at(t_ray *r, double t)
 {
 	t_vec3	tmp;
